+++ NEWS.md
@@ -6,6 +6,9 @@
 ### Fixed
 * `any_table()`: Fixed multi layerd column header labels where not applied correct. (28.11.2025)
 
+### GitHub related
+* Added article comparing this package with SAS even further. (02.12.2025)
+
 
 # qol 1.1.0 - CRAN release on 20.11.2025